+++ src/ui/about.rs
@@ -92,22 +92,10 @@ impl SimpleComponent for AboutDialog {
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
-                "<p>Added</p>",
-
-                "<ul>",
-                    "<li>Added `UpdatingPermissions` installation step</li>",
-                    "<li>Downloaders now will skip finished files and truncate them if needed</li>",
-                "</ul>",
-
                 "<p>Changed</p>",
 
                 "<ul>",
-                    "<li>Increased voiceovers version prediction error</li>",
-                    "<li>Updated FPS Unlocker version which fixes new game version integration issue</li>",
-                    "<li>Updated Turkish</li>",
-                    "<li>Updated German</li>",
-                    "<li>Updated Chinese</li>",
-                    "<li>Updated Polish</li>",
+                    "<li>Replaced FPS unlocker by a custom one</li>",
                 "</ul>"
             ].join("\n"),